First Major Commit
I have a map, with points... yay. Symbolizing is looking great, we have info window content. Paths for processing file need to be smarter... for some reason "./*filename" isn't working, though I assume thats on me.

I have the Isochrone stuff in there, but we are definitely going to need to integrate it effectively I see a lot of areas where this can break.